Alerts configured - Home 10% advanced

+++ frontend/src/actions/observations.ts
@@ -2,6 +2,7 @@ import { Dispatch } from "redux"
 import { clientAxios } from "../config/clientAxios"
 import { IObservation } from "../interfaces/interfaces";
 import { types } from "../types/types";
+import Swal from "sweetalert2";
 
 export const startGetObservations = () => {
     return async ( dispatch: Dispatch ) => {
@@ -22,7 +23,11 @@ export const startGetObservations = () => {
             })
             dispatch( setObservations( observations ) );
         } catch (error) {
-            console.log( error.response );
+            Swal.fire(
+              'Error',
+              error.response.data.msg,
+              'success'
+            )
         }
     }
 }
@@ -32,8 +37,17 @@ export const startCreateObservation = ( detail: string ) => {
         try {
             const vehicleId = getState().vehicles.activeVehicle.id;
             await clientAxios.post(`/observation/${ vehicleId }`, { detail });
+            Swal.fire(
+              'Observación agregada',
+              'La observación se agregó con éxito',
+              'success'
+            )
         } catch (error) {
-            console.log( error.response );
+            Swal.fire(
+              'Error',
+              error.response.data.msg,
+              'error'
+            )
         }
     }
 }
@@ -45,9 +59,40 @@ export const startSolveObservation = ( idState: number ) => {
             const { username } = getState().auth.user;
             const { data } = await clientAxios.patch( `observation/solve/${ observationId }`, { idState } );
             dispatch( updateStateObservation( data.observation.id, data.observation.idState, username ) )
+            Swal.fire(
+              'Observación actualizada',
+              'La observación se actualizó con éxito',
+              'success'
+            )
         } catch (error) {
-            console.log( error.response );
+            Swal.fire(
+              'Error',
+              error.response.data.msg,
+              'error'
+            )
         }
+    }
+}
+
+export const startModifyObservationDetail = ( detail: string ) => {
+    return async ( dispatch: Dispatch, getState: any ) => {
+        try {
+            const { id: observationId } = getState().observations.activeObservation;
+            const { data } = await clientAxios.patch(`/observation/${ observationId }`, { detail });
+            dispatch( modifyObservationDetail( observationId, data.observation.detail ) );
+            Swal.fire(
+              'Observación editada',
+              'La observación se editó con éxito',
+              'success'
+            )
+            
+        } catch (error) {
+            Swal.fire(
+              'Error',
+              error.response.data.msg,
+              'error'
+            )
+        }  
     }
 }
 
@@ -55,14 +100,33 @@ export const startDeleteObservation = () => {
     return async ( dispatch: Dispatch, getState: any ) => {
         try {
             const { id: observationId } = getState().observations.activeObservation;
-            const { data } = await clientAxios.delete(`/observation/${ observationId }`);
+            await clientAxios.delete(`/observation/${ observationId }`);
 
-            console.log( data );
-
-            dispatch( deleteObservation( observationId ) );
+            Swal.fire({
+              title: '¿Estás seguro?',
+              text: "No podrás recuperar la observación una vez eliminada",
+              icon: 'warning',
+              showCancelButton: true,
+              confirmButtonColor: '#3085d6',
+              cancelButtonColor: '#d33',
+              confirmButtonText: 'Sí, eliminar'
+            }).then((result) => {
+              if (result.isConfirmed) {
+                dispatch( deleteObservation( observationId ) );
+                Swal.fire(
+                  'Observación eliminada',
+                  'La observación se eliminó con éxito',
+                  'success'
+                )
+              }
+            })
 
         } catch (error) {
-            console.log( error.response );
+            Swal.fire(
+              'Error',
+              error.response.data.msg,
+              'error'
+            )
         }
     }
 }
@@ -86,7 +150,37 @@ export const updateStateObservation = ( id: number, idState: number, solver: str
     } }
 });
 
+export const startGetObservationsStatesPerUser = () => {
+    return async ( dispatch: Dispatch ) => {
+        try {
+            const { data } = await clientAxios.get('/observation/observation-per-user');
+            let observationsPerUser = [];
+            data.observations.map( ( info: any ) => {
+                // TODO
+            } )
+            dispatch( setObservationsStatesPerUser( data.observations ) );
+        } catch (error) {
+            console.log( error );
+        }  
+    }
+}
+
 export const deleteObservation = ( idToDelete: number ) => ({
     type: types.deleteObservation,
     payload: { idToDelete }
+})
+
+export const setWantEditDetail = ( bool: boolean ) => ({
+    type: types.wantEditDetail,
+    payload: { wantEditDetail: bool }
+})
+
+export const modifyObservationDetail = ( id: number, detail: string ) => ({
+    type: types.modifyObservationDetail,
+    payload: { observationToModify: { id, detail } }
+});
+
+export const setObservationsStatesPerUser = ( observationsQuantity: [] ) => ({
+    type: types.setObservationsStatesPerUser,
+    payload: observationsQuantity 
 })
